WFP-1922: revert to localstack 1.3.1

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppstier/client/TierToDeliusApiClient.kt
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.client.WebClient
 import org.springframework.web.reactive.function.client.awaitBody
+import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.ZonedDateTime
 
@@ -24,7 +25,7 @@ data class TierToDeliusResponse @JsonCreator constructor(
   val currentTier: String,
   val registrations: List<DeliusRegistration>,
   val convictions: List<DeliusConviction>,
-  val rsrscore: Int?,
+  val rsrscore: BigDecimal?,
   val ogrsscore: Int?,
 )
